move executor and collectors to the composer
