+++ crates/astria-composer/src/main.rs
@@ -51,7 +51,7 @@ async fn main() -> ExitCode {
 
     info!(config = cfg_ser, "initializing composer",);
 
-    Composer::from_config(&cfg)
+    let _ = Composer::from_config(&cfg)
         .await
         .expect("failed creating composer")
         .run_until_stopped()
