+++ crates/astria-composer/src/searcher/rollup.rs
@@ -6,13 +6,13 @@ use once_cell::sync::Lazy;
 use regex::Regex;
 
 #[derive(Debug)]
-pub(super) struct Rollup {
+pub(crate) struct Rollup {
     chain_id: String,
     url: String,
 }
 
 #[derive(Debug)]
-pub(super) struct ParseError {}
+pub(crate) struct ParseError {}
 
 impl ParseError {
     fn new() -> Self {
@@ -32,7 +32,7 @@ impl fmt::Display for ParseError {
 impl std::error::Error for ParseError {}
 
 impl Rollup {
-    pub(super) fn parse(from: &str) -> Result<Self, ParseError> {
+    pub(crate) fn parse(from: &str) -> Result<Self, ParseError> {
         static ROLLUP_RE: Lazy<Regex> = Lazy::new(|| {
             Regex::new(
                 r"(?x)
@@ -61,7 +61,7 @@ impl Rollup {
         })
     }
 
-    pub(super) fn into_parts(self) -> (String, String) {
+    pub(crate) fn into_parts(self) -> (String, String) {
         let Self {
             chain_id,
             url,
